+++ src/main/java/com/b1n_ry/yigd/config/YigdConfig.java
@@ -298,10 +298,10 @@ public class YigdConfig implements ConfigData {
         @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
         @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
         public DropRule defaultCuriosDropRule = DropRule.PUT_IN_GRAVE;
-//        public boolean enableBeansBackpacksCompat = true;
-//        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
-//        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
-//        public DropRule defaultBeansBackpacksDropRule = DropRule.PUT_IN_GRAVE;
+        public boolean enableCosmeticArmorCompat = true;
+        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
+        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
+        public DropRule defaultCosmeticArmorDropRule = DropRule.PUT_IN_GRAVE;
 //        public boolean enableRespawnObelisksCompat = true;
     }